remove ECOV test from codebase
