+++ R/utilitiesRVineEqualCorrOmega.R
@@ -834,138 +834,3 @@ omegaRvine = function(data, svcmDataFrame, indList, cPitData, theta, withRanks)
   return(omega)
 }
 
-
-getOmegaWithLikesCovE = function(data, svcmDataFrame, indList, cPitData, listOfMultipliers)
-{
-  
-  d <- ncol(data)
-  nObs = nrow(data)
-  nCopulas = d*(d-1)/2-1
-  
-  nParameters = sum(svcmDataFrame$nPar[1:nCopulas])
-  nGroups = length(indList)
-  
-  E = matrix(0,nrow=2*nGroups+nGroups,ncol=nParameters)
-  
-  for (jCopula in 1:nCopulas)
-  {
-    if (svcmDataFrame$nPar[jCopula])
-    {
-      family_1 = svcmDataFrame$family[jCopula]
-      par_1 = svcmDataFrame$par[[jCopula]]
-      if (jCopula<d)
-      {
-        cPit1_1 = data[,svcmDataFrame$var1[jCopula]]
-        cPit2_1 = data[,svcmDataFrame$var2[jCopula]]
-      }
-      else
-      {
-        cPit1_1 = getCpit1(cPitData, svcmDataFrame, jCopula)
-        cPit2_1 = getCpit2(cPitData, svcmDataFrame, jCopula)
-      }
-      
-      for (iGroup in 1:nGroups)
-      {
-        cPit1InGroup = cPit1_1[indList[[iGroup]]]
-        cPit2InGroup = cPit2_1[indList[[iGroup]]]
-        
-        nObsPerGroup = length(cPit1InGroup)
-        lambdaInGroup = nObsPerGroup/nObs
-        
-        E[1 + 2*(iGroup-1),svcmDataFrame$parInd[[jCopula]]] = deriv1LikeMult(par_1,cPit1InGroup,cPit2InGroup,family_1,listOfMultipliers$aInGroups[[iGroup]][,1])/lambdaInGroup 
-        E[2 + 2*(iGroup-1),svcmDataFrame$parInd[[jCopula]]] = deriv1LikeMult(par_1,cPit1InGroup,cPit2InGroup,family_1,listOfMultipliers$aInGroups[[iGroup]][,2])/lambdaInGroup
-        
-        E[(2*nGroups + iGroup),svcmDataFrame$parInd[[jCopula]]] = deriv1LikeMult(par_1,cPit1InGroup,cPit2InGroup,family_1,listOfMultipliers$bInGroups[[iGroup]])/lambdaInGroup
-      }
-    }
-  }
-  
-  
-  return(E)
-  
-  
-}
-
-
-omegaRvineCov = function(data, svcmDataFrame, indList, cPitData, theta, withRanks)
-{
-  
-  d <- ncol(data)
-  nObs = nrow(data)
-  nCopulas = d*(d-1)/2-1
-  
-  nParameters = sum(svcmDataFrame$nPar[1:nCopulas])
-  nParametersFirstTree = sum(svcmDataFrame$nPar[1:(d-1)])
-  
-  # Obtain the cPits to be tested
-  copulaInd = nrow(svcmDataFrame)
-  cPit1 = getCpit1(cPitData, svcmDataFrame, copulaInd)
-  cPit2 = getCpit2(cPitData, svcmDataFrame, copulaInd)
-  
-  # Obtain the subsamples
-  nGroups = length(indList)
-  
-  omega = matrix(0,nrow=nParameters+2*nGroups+nGroups,ncol=nParameters+2*nGroups+nGroups)
-  bb = matrix(0,nGroups,2*nGroups)
-  cc = matrix(0,nGroups,nGroups)
-  
-  aInGroups = vector("list",nGroups)
-  bInGroups = vector("list",nGroups)
-  
-  for (iGroup in 1:nGroups)
-  {
-    # Obtain the subsample
-    cPit1InGroup = cPit1[indList[[iGroup]]]
-    cPit2InGroup = cPit2[indList[[iGroup]]]
-    dataInGroup = data[indList[[iGroup]],]
-    
-    # Obtain the estimated parameters
-    mu1 = theta[(1 + 2*(iGroup-1))]
-    mu2 = theta[(2 + 2*(iGroup-1))]
-    
-    nObsPerGroup = length(cPit1InGroup)
-    lambdaInGroup = nObsPerGroup/nObs
-    
-    bInGroups[[iGroup]] = theta[(2*nGroups+iGroup)] - (cPit1InGroup - mu1) * (cPit2InGroup - mu2)
-    
-    aInGroups[[iGroup]] = cbind(mu1-cPit1InGroup,
-                                mu2-cPit2InGroup)
-    
-    omega[(nParameters+1 + 2*(iGroup-1)):(nParameters+2 + 2*(iGroup-1)),
-          (nParameters+1 + 2*(iGroup-1)):(nParameters+2 + 2*(iGroup-1))] = 1/nObsPerGroup *(t(aInGroups[[iGroup]]) %*% aInGroups[[iGroup]])/lambdaInGroup
-    
-    bb[iGroup,(1 + 2*(iGroup-1)):(2 + 2*(iGroup-1))] = 1/nObsPerGroup *(t(aInGroups[[iGroup]]) %*% bInGroups[[iGroup]])/lambdaInGroup
-    
-    cc[iGroup,iGroup] = mean(bInGroups[[iGroup]]^2)/lambdaInGroup
-    
-  }
-  
-  
-  omega[(nParameters+2*nGroups+1):(nParameters+2*nGroups+nGroups),(nParameters+1):(nParameters+2*nGroups)] = bb
-  
-  omega[(nParameters+2*nGroups+1):(nParameters+2*nGroups+nGroups),(nParameters+2*nGroups+1):(nParameters+2*nGroups+nGroups)] = cc
-  
-  
-  listOfMultipliers = list(aInGroups=aInGroups,bInGroups=bInGroups)
-  
-  if (nParameters)
-  {
-    omegaWithLikesD = getOmegaWithLikesD(data, svcmDataFrame, cPitData)
-    if(withRanks)
-    {
-      bSsp = bSspForCovWithRanks(data, svcmDataFrame, cPitData)
-      omegaWithLikesD = omegaWithLikesD + bSsp
-    }
-    omega[1:nParameters,1:nParameters] = omegaWithLikesD
-    
-    omega[(nParameters+1):(nParameters+2*nGroups+nGroups),1:nParameters] = getOmegaWithLikesCovE(data, svcmDataFrame, indList, cPitData, listOfMultipliers)
-  }
-  
-  
-  xx = t(omega)
-  omega[upper.tri(omega)] = xx[upper.tri(xx)]
-  
-  
-  return(omega)
-}
-
