+++ man/pacotestset.Rd
@@ -21,17 +21,17 @@ For \code{testType = CCC}:
 
 The grouping method which is used to obtain a partitioning of the support of the conditioning variable W.
 
-Possible values: \code{TreeCCC} | \code{SumMedian} | \code{SumThirdsI} | \code{SumThirdsII} | \code{SumThirdsIII} | \code{SumQuartiles} | \code{ProdMedian} | \code{ProdThirdsI} | \code{ProdThirdsII} | \code{ProdThirdsIII} | \code{ProdQuartiles} | \code{TreeEC} | \code{TreeECOV}
+Possible values: \code{TreeCCC} | \code{SumMedian} | \code{SumThirdsI} | \code{SumThirdsII} | \code{SumThirdsIII} | \code{SumQuartiles} | \code{ProdMedian} | \code{ProdThirdsI} | \code{ProdThirdsII} | \code{ProdThirdsIII} | \code{ProdQuartiles} | \code{TreeEC}
 }
 
 \item{expMinSampleSize}{
-For \code{testType = CCC} with \code{grouping = TreeCCC | TreeECOV | TreeEC}:
+For \code{testType = CCC} with \code{grouping = TreeCCC | TreeEC}:
 
 The minimum number of observations which are allocated to a group in the decision tree learning process. The default value is \code{100}.
 }
 
 \item{aggInfo}{
-For \code{testType = CCC} with \code{grouping = TreeCCC | TreeECOV | TreeEC}:
+For \code{testType = CCC} with \code{grouping = TreeCCC | TreeEC}:
 
 The method used for aggregating information in the conditioning set. The information in the conditioning set can be aggregated by either taking the mean of all variables or the pairwise mean. The result is added as an additional variable which can be used by the decision tree to partition the support of the conditioning variable W.
 
@@ -51,7 +51,7 @@ A logical variable indicating whether the asymptotic-variance covariance matrix 
 }
 
 \item{finalComparison}{
-For \code{testType = CCC} with \code{grouping = TreeCCC | TreeECOV | TreeEC}:
+For \code{testType = CCC} with \code{grouping = TreeCCC | TreeEC}:
 
 A variable specifying whether at the end of the decision tree all subsets being part of the partition are compared against each other or whether only the pair with the highest value of the test statistic is used.
 
@@ -59,13 +59,13 @@ Possible values: \code{pairwiseMax} | \code{all}
 }
 
 \item{penaltyParams}{
-For \code{testType = CCC} with \code{grouping = TreeCCC | TreeECOV | TreeEC}:
+For \code{testType = CCC} with \code{grouping = TreeCCC | TreeEC}:
 
 A vector of length two, specifying the functional form of the penalty. The penalty is a function of the sample size n and chosen to be lambda(n) = cn^(-beta). The first entry of the vector is specifying the level c of the penalty and needs to be a positive real number. The second entry of the vector is specifying the power beta of the penalty and needs to be chosen from the interval (0,1).
 }
 
 \item{gamma0Partition}{
-For \code{testType = CCC} with \code{grouping = TreeCCC | TreeECOV | TreeEC}:
+For \code{testType = CCC} with \code{grouping = TreeCCC | TreeEC}:
 
 The gamma0 partition. I.e., the partition which is favoured via the penalty under the H0.
